+++ src/cascadia/TerminalSettingsEditor/Microsoft.Terminal.Settings.Editor.vcxproj.filters
@@ -40,6 +40,7 @@
     <Page Include="Launch.xaml" />
     <Page Include="MainPage.xaml" />
     <Page Include="Profiles.xaml" />
+    <Page Include="Appearances.xaml" />
     <Page Include="Rendering.xaml" />
     <Page Include="Actions.xaml" />
     <Page Include="SettingContainerStyle.xaml" />
