+++ src/cascadia/TerminalSettingsEditor/Profiles.cpp
@@ -23,64 +23,6 @@ static const std::array<winrt::guid, 2> InBoxProfileGuids{
     winrt::guid{ 0x0caa0dad, 0x35be, 0x5f56, { 0xa8, 0xff, 0xaf, 0xce, 0xee, 0xaa, 0x61, 0x01 } } // Command Prompt
 };
 
-// Function Description:
-// - This function presents a File Open "common dialog" and returns its selected file asynchronously.
-// Parameters:
-// - customize: A lambda that receives an IFileDialog* to customize.
-// Return value:
-// (async) path to the selected item.
-template<typename TLambda>
-static winrt::Windows::Foundation::IAsyncOperation<winrt::hstring> OpenFilePicker(HWND parentHwnd, TLambda&& customize)
-{
-    auto fileDialog{ winrt::create_instance<IFileDialog>(CLSID_FileOpenDialog) };
-    DWORD flags{};
-    THROW_IF_FAILED(fileDialog->GetOptions(&flags));
-    THROW_IF_FAILED(fileDialog->SetOptions(flags | FOS_FORCEFILESYSTEM | FOS_NOCHANGEDIR | FOS_DONTADDTORECENT)); // filesystem objects only; no recent places
-    customize(fileDialog.get());
-
-    auto hr{ fileDialog->Show(parentHwnd) };
-    if (!SUCCEEDED(hr))
-    {
-        if (hr == HRESULT_FROM_WIN32(ERROR_CANCELLED))
-        {
-            co_return winrt::hstring{};
-        }
-        THROW_HR(hr);
-    }
-
-    winrt::com_ptr<IShellItem> result;
-    THROW_IF_FAILED(fileDialog->GetResult(result.put()));
-
-    wil::unique_cotaskmem_string filePath;
-    THROW_IF_FAILED(result->GetDisplayName(SIGDN_FILESYSPATH, &filePath));
-
-    co_return winrt::hstring{ filePath.get() };
-}
-
-// Function Description:
-// - Helper that opens a file picker pre-seeded with image file types.
-static winrt::Windows::Foundation::IAsyncOperation<winrt::hstring> OpenImagePicker(HWND parentHwnd)
-{
-    static constexpr COMDLG_FILTERSPEC supportedImageFileTypes[] = {
-        { L"All Supported Bitmap Types (*.jpg, *.jpeg, *.png, *.bmp, *.gif, *.tiff, *.ico)", L"*.jpg;*.jpeg;*.png;*.bmp;*.gif;*.tiff;*.ico" },
-        { L"All Files (*.*)", L"*.*" }
-    };
-
-    static constexpr winrt::guid clientGuidImagePicker{ 0x55675F54, 0x74A1, 0x4552, { 0xA3, 0x9D, 0x94, 0xAE, 0x85, 0xD8, 0xF2, 0x7A } };
-    return OpenFilePicker(parentHwnd, [](auto&& dialog) {
-        THROW_IF_FAILED(dialog->SetClientGuid(clientGuidImagePicker));
-        try
-        {
-            auto pictureFolderShellItem{ winrt::capture<IShellItem>(&SHGetKnownFolderItem, FOLDERID_PicturesLibrary, KF_FLAG_DEFAULT, nullptr) };
-            dialog->SetDefaultFolder(pictureFolderShellItem.get());
-        }
-        CATCH_LOG(); // non-fatal
-        THROW_IF_FAILED(dialog->SetFileTypes(ARRAYSIZE(supportedImageFileTypes), supportedImageFileTypes));
-        THROW_IF_FAILED(dialog->SetFileTypeIndex(1)); // the array is 1-indexed
-        THROW_IF_FAILED(dialog->SetDefaultExtension(L"jpg;jpeg;png;bmp;gif;tiff;ico"));
-    });
-}
-
 namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
 {
     Windows::Foundation::Collections::IObservableVector<Editor::Font> ProfileViewModel::_MonospaceFontList{ nullptr };
@@ -88,8 +30,8 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
 
     ProfileViewModel::ProfileViewModel(const Model::Profile& profile, const Model::CascadiaSettings& appSettings) :
         _profile{ profile },
+        _defaultAppearanceViewModel{ winrt::make<implementation::AppearanceViewModel>(profile.DefaultAppearance().try_as<AppearanceConfig>()) },
         _originalProfileGuid{ profile.Guid() },
-        _ShowAllFonts{ false },
         _appSettings{ appSettings }
     {
         // Add a property changed handler to our own property changed event.
@@ -97,17 +39,7 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
         //  unique view model members.
         PropertyChanged([this](auto&&, const PropertyChangedEventArgs& args) {
             const auto viewModelProperty{ args.PropertyName() };
-            if (viewModelProperty == L"BackgroundImagePath")
-            {
-                // notify listener that all background image related values might have changed
-                //
-                // We need to do this so if someone manually types "desktopWallpaper"
-                // into the path TextBox, we properly update the checkbox and stored
-                // _lastBgImagePath. Without this, then we'll permanently hide the text
-                // box, prevent it from ever being changed again.
-                _NotifyChanges(L"UseDesktopBGImage", L"BackgroundImageSettingsVisible");
-            }
-            else if (viewModelProperty == L"IsBaseLayer")
+            if (viewModelProperty == L"IsBaseLayer")
             {
                 // we _always_ want to show the background image settings in base layer
                 _NotifyChanges(L"BackgroundImageSettingsVisible");
@@ -118,24 +50,7 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
                 // NOTE: this is similar to what is done with BackgroundImagePath above
                 _NotifyChanges(L"UseParentProcessDirectory", L"UseCustomStartingDirectory");
             }
-            else if (viewModelProperty == L"FontFace")
-            {
-                // notify listener that all font face related values might have changed
-                if (!UsingMonospaceFont())
-                {
-                    _ShowAllFonts = true;
-                }
-                _NotifyChanges(L"ShowAllFonts", L"UsingMonospaceFont");
-            }
         });
-
-        // Cache the original BG image path. If the user clicks "Use desktop
-        // wallpaper", then un-checks it, this is the string we'll restore to
-        // them.
-        if (BackgroundImagePath() != L"desktopWallpaper")
-        {
-            _lastBgImagePath = BackgroundImagePath();
-        }
 
         // Do the same for the starting directory
         if (!StartingDirectory().empty())
@@ -291,41 +206,6 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
         return _MonospaceFontList;
     }
 
-    // Method Description:
-    // - Searches through our list of monospace fonts to determine if the settings model's current font face is a monospace font
-    // - NOTE: This is information stored from DWrite in _UpdateFontList()
-    bool ProfileViewModel::UsingMonospaceFont() const noexcept
-    {
-        bool result{ false };
-        const auto currentFont{ FontFace() };
-        for (const auto& font : _MonospaceFontList)
-        {
-            if (font.LocalizedName() == currentFont)
-            {
-                result = true;
-            }
-        }
-        return result;
-    }
-
-    // Method Description:
-    // - Determines whether we should show the list of all the fonts, or we should just show monospace fonts
-    bool ProfileViewModel::ShowAllFonts() const noexcept
-    {
-        // - _ShowAllFonts is directly bound to the checkbox. So this is the user set value.
-        // - If we are not using a monospace font, show all of the fonts so that the ComboBox is still properly bound
-        return _ShowAllFonts || !UsingMonospaceFont();
-    }
-
-    void ProfileViewModel::ShowAllFonts(const bool& value)
-    {
-        if (_ShowAllFonts != value)
-        {
-            _ShowAllFonts = value;
-            _NotifyChanges(L"ShowAllFonts");
-        }
-    }
-
     winrt::guid ProfileViewModel::OriginalProfileGuid() const noexcept
     {
         return _originalProfileGuid;
@@ -354,33 +234,9 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
         }
     }
 
-    bool ProfileViewModel::UseDesktopBGImage()
+    Editor::AppearanceViewModel ProfileViewModel::DefaultAppearance()
     {
-        return BackgroundImagePath() == L"desktopWallpaper";
-    }
-
-    void ProfileViewModel::UseDesktopBGImage(const bool useDesktop)
-    {
-        if (useDesktop)
-        {
-            // Stash the current value of BackgroundImagePath. If the user
-            // checks and un-checks the "Use desktop wallpaper" button, we want
-            // the path that we display in the text box to remain unchanged.
-            //
-            // Only stash this value if it's not the special "desktopWallpaper"
-            // value.
-            if (BackgroundImagePath() != L"desktopWallpaper")
-            {
-                _lastBgImagePath = BackgroundImagePath();
-            }
-            BackgroundImagePath(L"desktopWallpaper");
-        }
-        else
-        {
-            // Restore the path we had previously cached. This might be the
-            // empty string.
-            BackgroundImagePath(_lastBgImagePath);
-        }
+        return _defaultAppearanceViewModel;
     }
 
     bool ProfileViewModel::UseParentProcessDirectory()
@@ -429,11 +285,6 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
         }
     }
 
-    bool ProfileViewModel::BackgroundImageSettingsVisible()
-    {
-        return IsBaseLayer() || BackgroundImagePath() != L"";
-    }
-
     void ProfilePageNavigationState::DeleteProfile()
     {
         auto deleteProfileArgs{ winrt::make_self<DeleteProfileEventArgs>(_Profile.Guid()) };
@@ -441,86 +292,22 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
     }
 
     Profiles::Profiles() :
-        _ColorSchemeList{ single_threaded_observable_vector<ColorScheme>() },
         _previewControl{ Control::TermControl(Model::TerminalSettings{}, make<PreviewConnection>()) }
     {
         InitializeComponent();
 
-        INITIALIZE_BINDABLE_ENUM_SETTING(CursorShape, CursorStyle, winrt::Microsoft::Terminal::Core::CursorStyle, L"Profile_CursorShape", L"Content");
-        INITIALIZE_BINDABLE_ENUM_SETTING_REVERSE_ORDER(BackgroundImageStretchMode, BackgroundImageStretchMode, winrt::Windows::UI::Xaml::Media::Stretch, L"Profile_BackgroundImageStretchMode", L"Content");
         INITIALIZE_BINDABLE_ENUM_SETTING(AntiAliasingMode, TextAntialiasingMode, winrt::Microsoft::Terminal::Control::TextAntialiasingMode, L"Profile_AntialiasingMode", L"Content");
         INITIALIZE_BINDABLE_ENUM_SETTING_REVERSE_ORDER(CloseOnExitMode, CloseOnExitMode, winrt::Microsoft::Terminal::Settings::Model::CloseOnExitMode, L"Profile_CloseOnExit", L"Content");
         INITIALIZE_BINDABLE_ENUM_SETTING(ScrollState, ScrollbarState, winrt::Microsoft::Terminal::Control::ScrollbarState, L"Profile_ScrollbarVisibility", L"Content");
 
-        // manually add Custom FontWeight option. Don't add it to the Map
-        INITIALIZE_BINDABLE_ENUM_SETTING(FontWeight, FontWeight, uint16_t, L"Profile_FontWeight", L"Content");
-        _CustomFontWeight = winrt::make<EnumEntry>(RS_(L"Profile_FontWeightCustom/Content"), winrt::box_value<uint16_t>(0u));
-        _FontWeightList.Append(_CustomFontWeight);
-
-        // manually keep track of all the Background Image Alignment buttons
-        _BIAlignmentButtons.at(0) = BIAlign_TopLeft();
-        _BIAlignmentButtons.at(1) = BIAlign_Top();
-        _BIAlignmentButtons.at(2) = BIAlign_TopRight();
-        _BIAlignmentButtons.at(3) = BIAlign_Left();
-        _BIAlignmentButtons.at(4) = BIAlign_Center();
-        _BIAlignmentButtons.at(5) = BIAlign_Right();
-        _BIAlignmentButtons.at(6) = BIAlign_BottomLeft();
-        _BIAlignmentButtons.at(7) = BIAlign_Bottom();
-        _BIAlignmentButtons.at(8) = BIAlign_BottomRight();
-
-        // apply automation properties to more complex setting controls
-        for (const auto& biButton : _BIAlignmentButtons)
-        {
-            const auto tooltip{ ToolTipService::GetToolTip(biButton) };
-            Automation::AutomationProperties::SetName(biButton, unbox_value<hstring>(tooltip));
-        }
-
         const auto startingDirCheckboxTooltip{ ToolTipService::GetToolTip(StartingDirectoryUseParentCheckbox()) };
         Automation::AutomationProperties::SetFullDescription(StartingDirectoryUseParentCheckbox(), unbox_value<hstring>(startingDirCheckboxTooltip));
-
-        const auto backgroundImgCheckboxTooltip{ ToolTipService::GetToolTip(UseDesktopImageCheckBox()) };
-        Automation::AutomationProperties::SetFullDescription(UseDesktopImageCheckBox(), unbox_value<hstring>(backgroundImgCheckboxTooltip));
-
-        const auto showAllFontsCheckboxTooltip{ ToolTipService::GetToolTip(ShowAllFontsCheckbox()) };
-        Automation::AutomationProperties::SetFullDescription(ShowAllFontsCheckbox(), unbox_value<hstring>(showAllFontsCheckboxTooltip));
 
         Automation::AutomationProperties::SetName(DeleteButton(), RS_(L"Profile_DeleteButton/Text"));
 
         _previewControl.IsEnabled(false);
         _previewControl.AllowFocusWhenDisabled(false);
         ControlPreview().Child(_previewControl);
-    }
-
-    IInspectable Profiles::CurrentFontFace() const
-    {
-        // look for the current font in our shown list of fonts
-        const auto& profileVM{ State().Profile() };
-        const auto profileFontFace{ profileVM.FontFace() };
-        const auto& currentFontList{ profileVM.ShowAllFonts() ? profileVM.CompleteFontList() : profileVM.MonospaceFontList() };
-        IInspectable fallbackFont;
-        for (const auto& font : currentFontList)
-        {
-            if (font.LocalizedName() == profileFontFace)
-            {
-                return box_value(font);
-            }
-            else if (font.LocalizedName() == L"Cascadia Mono")
-            {
-                fallbackFont = box_value(font);
-            }
-        }
-
-        // we couldn't find the desired font, set to "Cascadia Mono" since that ships by default
-        return fallbackFont;
-    }
-
-    void Profiles::FontFace_SelectionChanged(IInspectable const& /*sender*/, SelectionChangedEventArgs const& e)
-    {
-        // NOTE: We need to hook up a selection changed event handler here instead of directly binding to the profile view model.
-        //       A two way binding to the view model causes an infinite loop because both combo boxes keep fighting over which one's right.
-        const auto selectedItem{ e.AddedItems().GetAt(0) };
-        const auto newFontFace{ unbox_value<Editor::Font>(selectedItem) };
-        State().Profile().FontFace(newFontFace.LocalizedName());
     }
 
     void Profiles::OnNavigatedTo(const NavigationEventArgs& e)
@@ -531,18 +318,6 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
         if (!_State.Profile().CompleteFontList() || !_State.Profile().MonospaceFontList())
         {
             ProfileViewModel::UpdateFontList();
-        }
-
-        const auto& colorSchemeMap{ _State.Schemes() };
-        for (const auto& pair : colorSchemeMap)
-        {
-            _ColorSchemeList.Append(pair.Value());
-        }
-
-        const auto& biAlignmentVal{ static_cast<int32_t>(_State.Profile().BackgroundImageAlignment()) };
-        for (const auto& biButton : _BIAlignmentButtons)
-        {
-            biButton.IsChecked(biButton.Tag().as<int32_t>() == biAlignmentVal);
         }
 
         // Set the text disclaimer for the text box
@@ -571,16 +346,7 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
         // and propagate those changes to the UI
         _ViewModelChangedRevoker = _State.Profile().PropertyChanged(winrt::auto_revoke, [=](auto&&, const PropertyChangedEventArgs& args) {
             const auto settingName{ args.PropertyName() };
-            if (settingName == L"CursorShape")
-            {
-                _PropertyChangedHandlers(*this, PropertyChangedEventArgs{ L"CurrentCursorShape" });
-                _PropertyChangedHandlers(*this, PropertyChangedEventArgs{ L"IsVintageCursor" });
-            }
-            else if (settingName == L"BackgroundImageStretchMode")
-            {
-                _PropertyChangedHandlers(*this, PropertyChangedEventArgs{ L"CurrentBackgroundImageStretchMode" });
-            }
-            else if (settingName == L"AntialiasingMode")
+            if (settingName == L"AntialiasingMode")
             {
                 _PropertyChangedHandlers(*this, PropertyChangedEventArgs{ L"CurrentAntiAliasingMode" });
             }
@@ -596,23 +362,13 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
             {
                 _PropertyChangedHandlers(*this, PropertyChangedEventArgs{ L"CurrentScrollState" });
             }
-            else if (settingName == L"FontWeight")
-            {
-                _PropertyChangedHandlers(*this, PropertyChangedEventArgs{ L"CurrentFontWeight" });
-                _PropertyChangedHandlers(*this, PropertyChangedEventArgs{ L"IsCustomFontWeight" });
-            }
-            else if (settingName == L"ColorSchemeName")
-            {
-                _PropertyChangedHandlers(*this, PropertyChangedEventArgs{ L"CurrentColorScheme" });
-            }
-            else if (settingName == L"FontFace" || settingName == L"CurrentFontList")
-            {
-                _PropertyChangedHandlers(*this, PropertyChangedEventArgs{ L"CurrentFontFace" });
-            }
-            else if (settingName == L"BackgroundImageAlignment")
-            {
-                _UpdateBIAlignmentControl(static_cast<int32_t>(_State.Profile().BackgroundImageAlignment()));
-            }
+            _previewControl.Settings(_State.Profile().TermSettings());
+            _previewControl.UpdateSettings();
+        });
+
+        // The Appearances object handles updating the values in the settings UI, but
+        // we still need to listen to the changes here just to update the preview control
+        _AppearanceViewModelChangedRevoker = _State.Profile().DefaultAppearance().PropertyChanged(winrt::auto_revoke, [=](auto&&, const PropertyChangedEventArgs& /*args*/) {
             _previewControl.Settings(_State.Profile().TermSettings());
             _previewControl.UpdateSettings();
         });
@@ -632,26 +388,7 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
     void Profiles::OnNavigatedFrom(const NavigationEventArgs& /*e*/)
     {
         _ViewModelChangedRevoker.revoke();
-    }
-
-    ColorScheme Profiles::CurrentColorScheme()
-    {
-        const auto schemeName{ _State.Profile().ColorSchemeName() };
-        if (const auto scheme{ _State.Schemes().TryLookup(schemeName) })
-        {
-            return scheme;
-        }
-        else
-        {
-            // This Profile points to a color scheme that was renamed or deleted.
-            // Fallback to Campbell.
-            return _State.Schemes().TryLookup(L"Campbell");
-        }
-    }
-
-    void Profiles::CurrentColorScheme(const ColorScheme& val)
-    {
-        _State.Profile().ColorSchemeName(val.Name());
+        _AppearanceViewModelChangedRevoker.revoke();
     }
 
     bool Profiles::IsBellStyleFlagSet(const uint32_t flag)
@@ -684,18 +421,6 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
     {
         auto state{ winrt::get_self<ProfilePageNavigationState>(_State) };
         state->DeleteProfile();
-    }
-
-    fire_and_forget Profiles::BackgroundImage_Click(IInspectable const&, RoutedEventArgs const&)
-    {
-        auto lifetime = get_strong();
-
-        const auto parentHwnd{ reinterpret_cast<HWND>(_State.WindowRoot().GetHostingWindow()) };
-        auto file = co_await OpenImagePicker(parentHwnd);
-        if (!file.empty())
-        {
-            _State.Profile().BackgroundImagePath(file);
-        }
     }
 
     fire_and_forget Profiles::Icon_Click(IInspectable const&, RoutedEventArgs const&)
@@ -765,76 +490,9 @@ namespace winrt::Microsoft::Terminal::Settings::Editor::implementation
         }
     }
 
-    IInspectable Profiles::CurrentFontWeight() const
-    {
-        // if no value was found, we have a custom value
-        const auto maybeEnumEntry{ _FontWeightMap.TryLookup(_State.Profile().FontWeight().Weight) };
-        return maybeEnumEntry ? maybeEnumEntry : _CustomFontWeight;
-    }
-
-    void Profiles::CurrentFontWeight(const IInspectable& enumEntry)
-    {
-        if (auto ee = enumEntry.try_as<Editor::EnumEntry>())
-        {
-            if (ee != _CustomFontWeight)
-            {
-                const auto weight{ winrt::unbox_value<uint16_t>(ee.EnumValue()) };
-                const Windows::UI::Text::FontWeight setting{ weight };
-                _State.Profile().FontWeight(setting);
-
-                // Profile does not have observable properties
-                // So the TwoWay binding doesn't update on the State --> Slider direction
-                FontWeightSlider().Value(weight);
-            }
-            _PropertyChangedHandlers(*this, PropertyChangedEventArgs{ L"IsCustomFontWeight" });
-        }
-    }
-
-    bool Profiles::IsCustomFontWeight()
-    {
-        // Use SelectedItem instead of CurrentFontWeight.
-        // CurrentFontWeight converts the Profile's value to the appropriate enum entry,
-        // whereas SelectedItem identifies which one was selected by the user.
-        return FontWeightComboBox().SelectedItem() == _CustomFontWeight;
-    }
-
-    void Profiles::BIAlignment_Click(IInspectable const& sender, RoutedEventArgs const& /*e*/)
-    {
-        if (const auto& button{ sender.try_as<Primitives::ToggleButton>() })
-        {
-            if (const auto& tag{ button.Tag().try_as<int32_t>() })
-            {
-                // Update the Profile's value and the control
-                _State.Profile().BackgroundImageAlignment(static_cast<ConvergedAlignment>(*tag));
-                _UpdateBIAlignmentControl(*tag);
-            }
-        }
-    }
-
-    // Method Description:
-    // - Resets all of the buttons to unchecked, and checks the one with the provided tag
-    // Arguments:
-    // - val - the background image alignment (ConvergedAlignment) that we want to represent in the control
-    void Profiles::_UpdateBIAlignmentControl(const int32_t val)
-    {
-        for (const auto& biButton : _BIAlignmentButtons)
-        {
-            if (const auto& biButtonAlignment{ biButton.Tag().try_as<int32_t>() })
-            {
-                biButton.IsChecked(biButtonAlignment == val);
-            }
-        }
-    }
-
-    bool Profiles::IsVintageCursor() const
-    {
-        return _State.Profile().CursorShape() == Core::CursorStyle::Vintage;
-    }
-
     void Profiles::Pivot_SelectionChanged(Windows::Foundation::IInspectable const& /*sender*/,
                                           RoutedEventArgs const& /*e*/)
     {
         _State.LastActivePivot(static_cast<Editor::ProfilesPivots>(ProfilesPivot().SelectedIndex()));
     }
-
 }
